Move common util => support/
