+++ spec/git_pair/usage_spec.rb
@@ -1,18 +1,5 @@
 require 'pivotal_git_scripts/git_pair'
-
-class MockGitConfig
-  include RSpec::Matchers
-  
-  def fun
-    ->(*args) do
-      @args = args
-    end
-  end
-
-  def must_have_received(expected={})
-    expect(@args).to include(expected)
-  end
-end
+require 'support'
 
 module UseCases
   class GitPair
